# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api for gophers Chat Project !",
        "title": "ChatX API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/chats/{chat_id}/messages": {
            "get": {
                "description": "Berilgan chat_id bo'yicha barcha xabarlar tarixini qaytaradi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Chat xabarlarini olish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat ID",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xabarlar ro'yxati: {\"data\": [MessageDetail ob'ektlari]}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/service.MessageDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/messages": {
            "post": {
                "description": "Yangi xabar yaratadi va uni guruh a'zolariga WebSocket orqali real-vaqtda tarqatadi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Xabar yuborish",
                "parameters": [
                    {
                        "description": "Xabar ma'lumotlari",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.Message"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Xabar yaratildi: {\"data\": {message_object}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/service.Message"
                            }
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri JSON formati",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages/{chat_id}": {
            "patch": {
                "description": "Chatdagi barcha xabarlarni joriy foydalanuvchi uchun o'qilgan holatiga o'tkazadi va bu haqda boshqa a'zolarga xabar beradi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Xabarlarni o'qilgan deb belgilash",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat ID",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli: {\"data\": {\"status\": \"success\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages/{id}": {
            "delete": {
                "description": "Xabarni ma'lumotlar bazasidan o'chiradi va WebSocket orqali barcha chat a'zolariga xabar o'chirilganligi haqida signal yuboradi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Xabarni o'chirish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "O'chirilishi kerak bo'lgan xabar IDsi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli: {\"data\": {\"result\": \"deleted\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Xabar topilmadi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Yuborilgan xabar matnini o'zgartiradi va bu haqda guruh a'zolariga WebSocket orqali xabar beradi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Xabarni tahrirlash",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Xabar IDsi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tahrirlash ma'lumotlari (message_text va chat_id)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli: {\"data\": {\"result\": \"updated\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri JSON yoki ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "description": "Tizimdagi foydalanuvchilarni qidirish va sahifalangan ro'yxatini qaytaradi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Foydalanuvchilar ro'yxatini olish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "So'rov yuborayotgan foydalanuvchi IDsi",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Natijalar soni",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Surilish (offset)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Qidiruv matni (ism yoki login)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchilar: {\"data\": [User ob'ektlari]}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/service.User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "service.Message": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer"
                },
                "chat_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_read": {
                    "type": "boolean"
                },
                "message_text": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "service.MessageDetail": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_name": {
                    "type": "string"
                }
            }
        },
        "service.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}